# openapi spec.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/url/content":{"get":{"summary":"Get Url Text","operationId":"get_url_text_url_content_get","parameters":[{"required":false,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/url/entities":{"get":{"summary":"Get Url Entities","operationId":"get_url_entities_url_entities_get","parameters":[{"required":false,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/url/entity-text":{"get":{"summary":"Get Url Entity Text","operationId":"get_url_entity_text_url_entity_text_get","parameters":[{"required":false,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"},{"required":false,"schema":{"title":"Entity","type":"string"},"name":"entity","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/add/":{"post":{"summary":"Add Urldata","operationId":"add_urldata_add__post","parameters":[{"required":false,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/database":{"get":{"summary":"View Database","operationId":"view_database_database_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/database/entities":{"get":{"summary":"Extract Entities","operationId":"extract_entities_database_entities_get","parameters":[{"required":true,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/database/entity-texts":{"get":{"summary":"Extract Entity Text","operationId":"extract_entity_text_database_entity_texts_get","parameters":[{"required":true,"schema":{"title":"Url","type":"string"},"name":"url","in":"query"},{"required":true,"schema":{"title":"Entity","type":"string"},"name":"entity","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload/":{"post":{"summary":"Add Data From File","description":"Future implementation, users can have the option of either uploading a file entering a URL","operationId":"add_data_from_file_upload__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_data_from_file_upload__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_add_data_from_file_upload__post":{"title":"Body_add_data_from_file_upload__post","required":["file"],"type":"object","properties":{"file":{"title":"File","type":"string","format":"binary"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}